// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        Booking: {
            /** Cancellation Reason */
            cancellation_reason?: string | null;
            /** Created At */
            created_at: string;
            /** Ends At */
            ends_at: string;
            /** Expires At */
            expires_at?: string | null;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Kind */
            kind: string;
            /**
             * Resource Id
             * Format: uuid
             */
            resource_id: string;
            /** Starts At */
            starts_at: string;
            /** Status */
            status: string;
            /** Updated At */
            updated_at: string;
            /** User Id */
            user_id?: string | null;
            /** Version */
            version: number;
        };
// ...
        Resource: {
            /** Active */
            active: boolean;
            /** Created At */
            created_at: string;
            /** Description */
            description: string;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Location */
            location: string;
            /** Name */
            name: string;
            /** Updated At */
            updated_at: string;
            /** Version */
            version: number;
        };
// ...
        User: {
            /** Created At */
            created_at: string;
            /** Display Name */
            display_name: string;
            /** Email */
            email: string;
            /** Enabled */
            enabled: boolean;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Role */
            role: string;
            /** Updated At */
            updated_at: string;
            /** Version */
            version: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}